backend/requirements.txt
alembic==1.6.5
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
gql==2.0.0
graphql-core==2.3.2
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
promise==2.3
protobuf==3.17.3
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
requests==2.25.1
Rx==1.6.1
six==1.16.0
SQLAlchemy==1.4.19
urllib3==1.26.6
Werkzeug==2.0.1